requirements.txt
omegaconf==2.1.2
timm==0.8.3.dev0
numpy
opencv-python
pycocotools
lxml
tqdm
pillow
h5py
pandas
scipy
wandb
transformers
matplotlib
wilds
gdown
tabulate
yacs
scikit-learn
einops
pytorch_lightning
git+https://github.com/openai/CLIP.git